Added run_with_hydro_daily
run with hydro_daily function is added which is similar to `run_with_hydro` function but is able to output fluxes on daily timescale. For this a mass balance model which has a function `get_daily_mb`is necessary which is available in TIModel of massbalance sandbox and not in OGGM core.

compile_run_output had to be adjusted to compile the output of the run_with_hydro_daily function in the format (years,days)